dom/plugins/test/mochitest/test_bug985859.html

<!doctype html>
<html>
<head>
  <title>Test for Bug 985859</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="plugin-utils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <meta http-equiv="content-type" content="text/html; charset=utf-8">
  <base href="chrome://browser/content/">
</head>
<body>
<script type="application/javascript">

setTestPluginEnabledState(SpecialPowers.Ci.nsIPluginTag.STATE_ENABLED);

var testType     = navigator.mimeTypes["application/x-test"];
var testTypeCap  = navigator.mimeTypes["Application/x-Test"];
var testTypeCap2 = navigator.mimeTypes["APPLICATION/X-TEST"];

ok(testType, "Test plugin should be found");
is(testType, testTypeCap, "navigator.mimeTypes should be case insensitive");
is(testType, testTypeCap2, "navigator.mimeTypes should be case insensitive");

</script>
</body>
</html>
